remove GUI & refactor the rest of the code
